+++ pype/modules/clockify/__init__.py
@@ -1,4 +1,4 @@
-from .clockify import ClockifyModule
+from .clockify_module import ClockifyModule
 
 __all__ = (
     "ClockifyModule",
